Workaround chicken-and-egg dep problem with setup.py
setup.py was trying to import the signalfx module to get to the name and
version, which triggered the code to be evaluated, which requires the
dependencies to be installed. This obviously is a chicken-and-egg
problem that prevents from installing the package correctly.

Workaround the problem by using execfile() instead to get the name and
version into the global scope without importing the signalfx module.