Major rework of the SignalFlow client to support multiple transports
Refactor the SignalFlow client code to support multiple client and
expose the full surface of the SignalFlow 2.0 API (even if server
support is not complete yet).

The new structure allows for implementing various transports to interact
with the SignalFlow API, in particular to support both the SSE-over-HTTP
endpoints and the WebSocket endpoint.

Reconnection and re-issue of execute/attach request is automatic on loss
of connection and is transparently handled by the Computation object
with the collaboration of the underlying transport.