Wait for authenticated message before sending other requests
Fixes a potential race condition in which we would send `execute`
requests (or others) before the server had processed our authentication
request, resulting in the server closing the connection. The WebSocket
endpoint now sends an authentication confirmation message that we can
wait for.

Also improve error handling related to authentication / establishing the
connection.