SignalFx Python client library version 1.0.0
Major refactoring of the structure of the library to split between the
REST API client, the Ingest client and the SignalFlow client. Now
getting a SignalFx object gets a thin base that gives access to the
three API sub-clients.

The appropriate token needs to be passed when retrieving one of those
sub-clients since different tokens must be used (ingest requires
organization or team level access tokens, while the REST and SignalFlow
API require org-scoped user tokens).

Since this is a major change, and because of the release of the
SignalFlow API and its client, we're bumping the version to 1.x.