Immediately release data messages when ready
The previous implementation would only yield a data message when the
timestamp changed (signaling the next series of data batches to
compose). Now we figure out the expected number of data batches per
composed data message and as soon as we have all of them for a
timestamp, we can yield the data message immediately.

This effectively reduces the perceived data lag by one resolution.