Make validate_detector return Response object
There is valuable information about why the detector model was not
validated inside of resp.text. Validation can fail because of an
invalid token or the SignalFlow syntax. Raising an exception inside
validate_detector currently conceals this information by setting
resp.content and resp.text but not updating resp.reason which is
displayed by calling resp.raise_for_status.

Let the user decide if they want to raise_for_status() and leave it to
them how to set resp.reason.