signalflow: add support for immediate computations
When immediate is set to true, the stop time and maximum delay of the
computation will be automatically adjusted to allow for the immediate
delivery of the computation output.

Test Plan: wrote a small python script using the library and verified
the setting was picked up. This is a non breaking change.